sensord: merge tizen 2.3 sensord into tizen branch
there are the quite many changes between private and public.
but almost patches are for code-clean and mobile specific features.
through updating sensord at once, private, tizen 2.3, tizen 3.0 code are
in sync.

+++ src/rotation_vector/rv/rv_sensor.cpp
@@ -203,13 +203,13 @@ bool rv_sensor::on_start(void)
 	AUTOLOCK(m_mutex);
 
 	m_accel_sensor->add_client(ACCELEROMETER_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_accel_sensor->add_interval((int)this, (m_interval/MS_TO_US), false);
+	m_accel_sensor->add_interval((intptr_t)this, (m_interval/MS_TO_US), false);
 	m_accel_sensor->start();
 	m_gyro_sensor->add_client(GYROSCOPE_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_gyro_sensor->add_interval((int)this, (m_interval/MS_TO_US), false);
+	m_gyro_sensor->add_interval((intptr_t)this, (m_interval/MS_TO_US), false);
 	m_gyro_sensor->start();
 	m_magnetic_sensor->add_client(GEOMAGNETIC_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_magnetic_sensor->add_interval((int)this, (m_interval/MS_TO_US), false);
+	m_magnetic_sensor->add_interval((intptr_t)this, (m_interval/MS_TO_US), false);
 	m_magnetic_sensor->start();
 
 	activate();
@@ -221,13 +221,13 @@ bool rv_sensor::on_stop(void)
 	AUTOLOCK(m_mutex);
 
 	m_accel_sensor->delete_client(ACCELEROMETER_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_accel_sensor->delete_interval((int)this, false);
+	m_accel_sensor->delete_interval((intptr_t)this, false);
 	m_accel_sensor->stop();
 	m_gyro_sensor->delete_client(GYROSCOPE_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_gyro_sensor->delete_interval((int)this, false);
+	m_gyro_sensor->delete_interval((intptr_t)this, false);
 	m_gyro_sensor->stop();
 	m_magnetic_sensor->delete_client(GEOMAGNETIC_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_magnetic_sensor->delete_interval((int)this, false);
+	m_magnetic_sensor->delete_interval((intptr_t)this, false);
 	m_magnetic_sensor->stop();
 
 	deactivate();
